+++ examples/simple.rs
@@ -371,7 +371,6 @@ impl SimpleWindow {
 
     fn add_hello_world(&mut self) {
         let mut text_widget = TextWidget::new(
-            self.pushrod.borrow_mut().get_factory(),
             "assets/OpenSans-Regular.ttf".to_string(),
             "Welcome to rust-pushrod!".to_string(),
             32,
@@ -401,7 +400,6 @@ impl SimpleWindow {
             .add_widget("BaseWidget1", Box::new(base_widget));
 
         let mut button1 = PushButtonWidget::new(
-            self.pushrod.borrow_mut().get_factory(),
             "assets/OpenSans-Regular.ttf".to_string(),
             "Random Color".to_string(),
             18,
@@ -420,7 +418,6 @@ impl SimpleWindow {
         );
 
         let mut button2 = PushButtonWidget::new(
-            self.pushrod.borrow_mut().get_factory(),
             "assets/OpenSans-Regular.ttf".to_string(),
             "Hide".to_string(),
             18,
@@ -438,7 +435,6 @@ impl SimpleWindow {
             .add_widget("HideButton1", Box::new(button2));
 
         let mut button3 = PushButtonWidget::new(
-            self.pushrod.borrow_mut().get_factory(),
             "assets/OpenSans-Regular.ttf".to_string(),
             "Disable".to_string(),
             18,
@@ -470,7 +466,6 @@ impl SimpleWindow {
             .add_widget("BoxWidget1", Box::new(box_widget));
 
         let mut text_widget2 = TextWidget::new(
-            self.pushrod.borrow_mut().get_factory(),
             "assets/OpenSans-Regular.ttf".to_string(),
             "Left".to_string(),
             24,
@@ -486,7 +481,6 @@ impl SimpleWindow {
         );
 
         let mut text_widget3 = TextWidget::new(
-            self.pushrod.borrow_mut().get_factory(),
             "assets/OpenSans-Regular.ttf".to_string(),
             "Center".to_string(),
             24,
@@ -502,7 +496,6 @@ impl SimpleWindow {
         );
 
         let mut text_widget4 = TextWidget::new(
-            self.pushrod.borrow_mut().get_factory(),
             "assets/OpenSans-Regular.ttf".to_string(),
             "Right".to_string(),
             24,
@@ -518,7 +511,6 @@ impl SimpleWindow {
         );
 
         let mut button2 = PushButtonWidget::new(
-            self.pushrod.borrow_mut().get_factory(),
             "assets/OpenSans-Regular.ttf".to_string(),
             "Hide".to_string(),
             18,
@@ -587,7 +579,6 @@ impl SimpleWindow {
             .add_widget_to_parent("Box5", Box::new(inner_box_4), box_1_id);
 
         let mut button = PushButtonWidget::new(
-            self.pushrod.borrow_mut().get_factory(),
             "assets/OpenSans-Regular.ttf".to_string(),
             "Hide".to_string(),
             18,
@@ -607,7 +598,6 @@ impl SimpleWindow {
 
     fn add_powered_by(&mut self) {
         let mut image_widget = ImageButtonWidget::new(
-            self.pushrod.borrow_mut().get_factory(),
             "assets/OpenSans-Regular.ttf".to_string(),
             "Powered By Rust!".to_string(),
             "rust-512x512.jpg".to_string(),
@@ -639,7 +629,6 @@ impl SimpleWindow {
             .add_widget("ProgressWidget", Box::new(progress_widget));
 
         let mut radio_1 = RadioButtonWidget::new(
-            self.pushrod.borrow_mut().get_factory(),
             "assets/OpenSans-Regular.ttf".to_string(),
             "1".to_string(),
             20,
@@ -657,7 +646,6 @@ impl SimpleWindow {
             .add_widget("Radio1", Box::new(radio_1));
 
         let mut radio_2 = RadioButtonWidget::new(
-            self.pushrod.borrow_mut().get_factory(),
             "assets/OpenSans-Regular.ttf".to_string(),
             "2".to_string(),
             20,
@@ -675,7 +663,6 @@ impl SimpleWindow {
             .add_widget("Radio2", Box::new(radio_2));
 
         let mut radio_3 = RadioButtonWidget::new(
-            self.pushrod.borrow_mut().get_factory(),
             "assets/OpenSans-Regular.ttf".to_string(),
             "3".to_string(),
             20,
@@ -693,7 +680,6 @@ impl SimpleWindow {
             .add_widget("Radio3", Box::new(radio_3));
 
         let mut progress_text = TextWidget::new(
-            self.pushrod.borrow_mut().get_factory(),
             "assets/OpenSans-Regular.ttf".to_string(),
             "50%".to_string(),
             18,
@@ -708,7 +694,6 @@ impl SimpleWindow {
             .add_widget("ProgressText1", Box::new(progress_text));
 
         let mut button1 = ToggleButtonWidget::new(
-            self.pushrod.borrow_mut().get_factory(),
             "assets/OpenSans-Regular.ttf".to_string(),
             "Animate".to_string(),
             18,
@@ -727,7 +712,6 @@ impl SimpleWindow {
             .add_widget("AnimateButton1", Box::new(button1));
 
         let mut button2 = PushButtonWidget::new(
-            self.pushrod.borrow_mut().get_factory(),
             "assets/OpenSans-Regular.ttf".to_string(),
             "Randomize".to_string(),
             18,
@@ -757,7 +741,6 @@ impl SimpleWindow {
 
     fn add_debugging(&mut self) {
         let mut check_widget = CheckboxWidget::new(
-            self.pushrod.borrow_mut().get_factory(),
             "assets/OpenSans-Regular.ttf".to_string(),
             "Enable Debugging".to_string(),
             20,
@@ -772,7 +755,6 @@ impl SimpleWindow {
             .add_widget("DebugCheck1", Box::new(check_widget));
 
         let mut text_widget1 = TextWidget::new(
-            self.pushrod.borrow_mut().get_factory(),
             "assets/OpenSans-Regular.ttf".to_string(),
             "Current Widget: 0".to_string(),
             20,
@@ -786,7 +768,6 @@ impl SimpleWindow {
             .add_widget("DebugText1", Box::new(text_widget1));
 
         let mut text_widget2 = TextWidget::new(
-            self.pushrod.borrow_mut().get_factory(),
             "assets/OpenSans-Regular.ttf".to_string(),
             "".to_string(),
             20,