+++ src/widget/push_button_widget.rs
@@ -37,7 +37,6 @@ impl PushButtonWidget {
     /// and the font justification when rendered.  Fonts are loaded from the `assets/`
     /// directory.
     pub fn new(
-        factory: &mut GfxFactory,
         font_name: String,
         text: String,
         font_size: u32,
@@ -47,7 +46,6 @@ impl PushButtonWidget {
             config: Configurable::new(),
             base_widget: BoxWidget::new(),
             text_widget: TextWidget::new(
-                factory,
                 font_name.to_string(),
                 text.to_string(),
                 font_size,